split ftp and smtp servers
